requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9.0
nuscenes-devkit>=1.1.10
numpy>=1.23
Pillow>=9.0
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.66
wandb>=0.16
pyquaternion>=0.9.9
